refactor intervals
